serverless/requirements.txt
azure-core==1.30.0
azure-storage-blob==12.19.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pycparser==2.21
requests==2.31.0
six==1.16.0
typing-extensions==4.9.0
urllib3==2.2.0
werkzeug==3.0.1
zipp==3.17.0
